# Optimize SQLite performance