This PR will optimize the performance of SQLite in ramfs and ext2:
1. SQLite-ramfs: Performance improves from 3.759s to 2.201s, reaching 94% of Linux performance.
2. SQLite-ext2: Performance improves from 4.839s to 3.376s, reaching 97.6% of Linux performance.